refactored state to be om message list page, moved functions to message list page
